refactored Entity classes folder structure
